requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.2.1